Suggestions about @Repository annotation usage
Hi, I found that there may be some minor improvements about annotations in your code.

It would be more appropriate to add @repository annotation at the class level. The @repository annotation indicates that the class provides the mechanism for storage, retrieval, search, update and delete operation on objects.

For better understanding and maintainability of a large application, adding @repository is a better choice.

Hope my suggestions are useful!